Added add info script that will allow you to add info about school qualification (more support coming soon)
The data must be passed as post data, it must have the user logged in and it must contain the following data: type, school_name, grad_year, qualification, t_subjects, subject_name_1, subject_score_1 (subject_name_2 and so on). Please check addQualification.html for full info.

form is in : addQualification.html

form handler is in : qualificationForm.js

The db has been updated to support school qualifications.

The following is not supported yet, but is planned:
1. Front-end Display of the qualification after adding the qualifiaction
2. Back-end API to get all the school university detail about a specific user
3. Back-end validaition for addInfo.php of type school qualifications (frontend is optional)